+++ event-service/src/main/java/com/fred/event_service/model/EventImage.java
@@ -17,8 +17,8 @@ import lombok.Setter;
 public class EventImage {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
-    private Integer eventId;
+    private Long id;
+    private Long eventId;
     private String imageUri;
     private Boolean isMain;
 }
